refactor(ui): move node menu to separate component
